Imprint

+++ routes/web.php
@@ -16,7 +16,11 @@ use App\Http\Controllers\TranslateController;
 */
 
 Route::get('/', function () {
-    return view('start', ['languages' => TranslateController::$languages, 'input' => '', 'searchCategory' => '', 'toLanguage' => 'enUs']);
+    return view('start', ['languages' => TranslateController::$languages, 'input' => '', 'searchCategory' => '', 'toLanguage' => 'enUS']);
+});
+
+Route::get('/imprint', function () {
+    return view('imprint');
 });
 
 Route::middleware('auth')->group(function () {
@@ -28,14 +32,16 @@ Route::get('/translate', function () {
 });
 Route::post('/translate', [TranslateController::class, 'translate']);
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+if (App::environment(['local']) === true) {
+    Route::get('/dashboard', function () {
+        return view('dashboard');
+    })->middleware(['auth', 'verified'])->name('dashboard');
 
-Route::middleware('auth')->group(function () {
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-});
+    Route::middleware('auth')->group(function () {
+        Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
+        Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
+        Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+    });
 
-require __DIR__.'/auth.php';
+    require __DIR__ . '/auth.php';
+}
